AI-Chatbot-Cloud-Storage/requirements.txt
streamlit
google-generativeai
python-dotenv
supabase